rename - testing tasks, moving testing process_metadata to test_process_metadata in scripts dir
